requirements.txt
google-generativeai>=0.8.0
opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=1.26.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
boto3>=1.35.0
